feat: decompile 11 functions (RSBS mask-clear+call wrappers, bit-test conditional, switch dispatchers, conditional returns, do-while loops, D_03004004 indexed hw writes, multi-call scene wrappers)

+++ src/decomp/asm_080143a0.c
@@ -2,7 +2,7 @@
 #include "global.h"
 #include "scenes.h"
 
-extern void func_0801429C(u8, u32);
+extern void func_0801429C(u32, u32);
 extern void func_08014374(void);
 
 void func_080143A0(void) {
